+++ CHANGELOG.md
@@ -22,6 +22,12 @@ All notable changes to this project will be documented here. For more details, v
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## UNRELEASED
+
+### Fixed
+
+- Request download URL only when the download is about to begin, to prevent `403 Forbidden` errors due to expiration (Filester)
+
 ## [10.0.1] - 2026-06-29
 
 ### Changed